requirements.txt
networkx
requests
praw
python-dotenv
streamlit==1.25.0